requirements.txt
branca==0.7.0
Django==4.2
earthengine_api==0.1.378
folium==0.15.0
keras==2.14.0
matplotlib==3.7.1
numpy
opencv_python==4.7.0.72
patchify==0.2.3
Pillow
protobuf==4.25.1
Requests==2.31.0
scikit_learn==1.2.2
segmentation_models
tensorflow
tensorflow_intel

